Table + Minimal styling completed
Elected to go for a vanilla HTML + CSS table rather than an Angular material table, as a material table would require further JSON parsing and transformation. A material table also doesn't do anything that can't be achieved with CSS, and doesn't justify using more memory and processor time to prepare data to be fed into it.